\name{htmlWidgetOutput}
\alias{htmlWidgetOutput}
\title{Create an arbitrary HTML widget}
\usage{
  htmlWidgetOutput(outputId, ...)
}
\arguments{
  \item{outputId}{output variable which will be used in
  addressing from the server.}

  \item{...}{Additional elements to be included in the
  widget.}
}
\description{
  Create an arbitrary HTML widget
}
\author{
  Jeff Allen <jeff.allen@trestletechnology.net>
}

