\name{gridsterItem}
\alias{gridsterItem}
\title{Add a Gridster item within a Gridster frame to a Shiny web page}
\usage{
  gridsterItem(..., col = NULL, row = NULL, size.x = NULL,
    size.y = NULL)
}
\arguments{
  \item{col}{The column in which to put the item.}

  \item{row}{The row in which to put the item.}

  \item{size.x}{Width, in number of tiles.}

  \item{size.y}{Height, in number of tiles.}

  \item{...}{Other properties or elements to include.}
}
\description{
  Add a Gridster item within a Gridster frame to a Shiny
  web page
}
\author{
  Winston Chang
}
\seealso{
  \code{\link{gridster}}
}

